added a navigation drawer and moved the menu
