+++ app/src/main/res/drawable/side_nav_bar.xml
@@ -1,0 +1,8 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="135"
+        android:endColor="#81c9c4"
+        android:startColor="#579794"
+        android:type="linear" />
+</shape>
