+++ app/src/main/res/layout/content_main.xml
@@ -33,7 +33,7 @@
         android:padding="@dimen/activity_horizontal_margin"
         android:background="@null"
         android:text="@string/start_label"
-        android:textSize="16sp"
+        android:textSize="13sp"
         android:textColor="@color/lightTransparent"
         android:fontFamily="sans-serif-light"
         android:visibility="visible"
@@ -48,7 +48,7 @@
         android:padding="@dimen/activity_horizontal_margin"
         android:background="@null"
         android:text="@string/stop_label"
-        android:textSize="16sp"
+        android:textSize="13sp"
         android:textColor="@color/lightTransparent"
         android:fontFamily="sans-serif-light"
         android:visibility="invisible"/>
