+++ app/src/main/res/values/resources.xml
@@ -16,4 +16,8 @@
         <item>@string/donation_3_dollars</item>
         <item>@string/donation_5_dollars</item>
     </string-array>
+
+    <string name="navigation_drawer_open" translatable="false">Open navigation drawer</string>
+    <string name="navigation_drawer_close" translatable="false">Close navigation drawer</string>
+
 </resources>
